chore: make lsp a subcommand of toto cli
