+++ crates/toto_cli/src/main.rs
@@ -1,7 +1,6 @@
-use std::{env, process::exit};
+use std::{env, error::Error};
 
 use clap::{Parser, Subcommand};
-use colored::Colorize;
 
 mod models;
 use toto_parser::{get_errors, report_error};
@@ -17,18 +16,32 @@ struct Args {
 #[derive(Subcommand, Debug)]
 #[command(version, about, long_about = None)]
 enum Command {
+    /// lint TOSCA file
+    ///
+    /// This command will report any grammar issues it can find
+    /// within provided file and imported files
     Check { path: String },
+
+    /// start language server
+    ///
+    /// This command will start a TOSCA language server, which
+    /// will provide various services to support the development
+    /// and validation of TOSCA files. This includes autocompletion,
+    /// real-time error checking, enhancing the efficiency and accuracy
+    /// of working with TOSCA files
+    LS,
 }
 
-fn main() {
+fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
     let args = Args::parse();
 
     match args.command {
         Command::Check { path } => check(path),
+        Command::LS => run_ls(),
     }
 }
 
-fn check(path: String) {
+fn check(path: String) -> Result<(), Box<dyn Error + Send + Sync>> {
     let mut ast = toto_ast::AST::<models::Entity, models::Relation>::new();
 
     let doc_path = "file://".to_string() + env::current_dir().unwrap().to_str().unwrap() + "/";
@@ -46,8 +59,13 @@ fn check(path: String) {
         .for_each(|(what, loc)| report_error(what, loc, &ast));
 
     if !has_errors {
-        println!("{}", "valid!".green());
+        Ok(())
+    } else {
+        Err("".into())
     }
+}
 
-    exit(if has_errors { 1 } else { 0 });
+fn run_ls() -> Result<(), Box<dyn Error + Send + Sync>> {
+    let server = toto_lsp::server::Server::new();
+    server.run()
 }
