+++ crates/toto_lsp/src/capabilities/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod complete;
+pub(crate) mod diagnostics;
+pub(crate) mod goto_definition;
