+++ crates/toto_lsp/src/lib.rs
@@ -1,0 +1,3 @@
+mod capabilities;
+mod models;
+pub mod server;
